Refactor REDCapConnection class
