+++ common/src/python/redcap/redcap_connection.py
@@ -1,10 +1,11 @@
-"""Classes and methods for connecting to REDCap."""
+"""Classes and methods for connecting to REDCap via API."""
 import json
 from json import JSONDecodeError
 from typing import Any, Dict, List, Optional
 
 import requests
 from inputs.parameter_store import REDCapParameters, REDCapReportParameters
+from redcap.redcap_project import REDCapProject
 from requests import Response
 
 
@@ -89,9 +90,10 @@ class REDCapSuperUserConnection:
 
 
 class REDCapConnection:
-    """Class managing the connection to a REDCap project.
+    """Class managing the API requests to a REDCap project.
 
-    Provides a post method to adapting classes. See `ProjectReader`
+    Provides a post method to adapting classes. See
+    `REDCapReportConnection`
     """
 
     def __init__(self, *, token: str, url: str) -> None:
@@ -104,7 +106,6 @@ class REDCapConnection:
         """
         self.__token = token
         self.__url = url
-        self.__set_project_info()
 
     @classmethod
     def create_from(cls, parameters: REDCapParameters) -> 'REDCapConnection':
@@ -119,46 +120,26 @@ class REDCapConnection:
         return REDCapConnection(token=parameters['token'],
                                 url=parameters['url'])
 
-    @property
-    def pid(self) -> int:
-        """Returns REDCap project ID."""
-        return self.__pid
+    def get_project(self) -> 'REDCapProject':
+        """Get the REDCap project for this connection."""
 
-    @property
-    def title(self) -> str:
-        """Returns REDCap project title."""
-        return self.__title
+        project_info = self.__export_project_info()
+        field_names = self.__export_field_names()
 
-    @property
-    def primary_key_field(self) -> str:
-        """Returns primary key field for the project."""
-        return self.__pk_field
-
-    def __set_project_info(self):
-        """Set project attributes and primary key field."""
-
-        project_info = self.export_project_info()
-        self.__pid = int(project_info['project_id'])
-        self.__title = project_info['project_title']
-        self.__longitudinal = (project_info['is_longitudinal'] == 1)
-        self.__repeating_ins = (
-            project_info['has_repeating_instruments_or_events'] == 1)
-        field_names = self.export_field_names()
-        self.__pk_field = field_names[0]['export_field_name']
-
-    def is_longitudinal(self) -> bool:
-        """Is this project set up to collect longitudinal data."""
-        return self.__longitudinal
-
-    def has_repeating_instruments_or_events(self) -> bool:
-        """Does this project has repeating instruments or events."""
-        return self.__repeating_ins
+        return REDCapProject(
+            redcap_con=self,
+            pid=int(project_info['project_id']),
+            title=project_info['project_title'],
+            pk_field=field_names[0]['export_field_name'],
+            longitudinal=(project_info['is_longitudinal'] == 1),
+            repeating_ins=(
+                project_info['has_repeating_instruments_or_events'] == 1))
 
     def post_request(self,
                      *,
                      data: Dict[str, str],
                      result_format: Optional[str] = None,
-                     return_format: str = 'json') -> Any:
+                     error_format: str = 'json') -> Any:
         """Posts a request to the REDCap project with the given data object.
 
         Returns:
@@ -168,7 +149,7 @@ class REDCapConnection:
           REDCapConnectionError if there is an error connecting to the
           specified project
         """
-        data.update({'token': self.__token, 'returnFormat': return_format})
+        data.update({'token': self.__token, 'returnFormat': error_format})
         if result_format:
             data['format'] = result_format
         try:
@@ -221,100 +202,7 @@ class REDCapConnection:
 
         return response.text
 
-    def import_records(self, records: str, data_format: str = 'json') -> int:
-        """Import records to the REDCap project.
-
-        Args:
-            records: List of records to be imported as a csv/json string
-            data_format (optional): Import formart, defaults to 'json'.
-
-        Raises:
-          REDCapConnectionError if the response has an error.
-        """
-
-        message = "importing records"
-        data = {
-            'content': 'record',
-            'action': 'import',
-            'forceAutoNumber': 'false',
-            'data': records,
-            'returnContent': 'count',
-        }
-
-        response = self.post_request(data=data, result_format=data_format)
-        if not response.ok:
-            raise REDCapConnectionError(
-                message=error_message(message=message, response=response))
-
-        try:
-            num_records = json.loads(response.text)['count']
-        except (JSONDecodeError, ValueError) as error:
-            raise REDCapConnectionError(message=message) from error
-
-        return num_records
-
-    def export_records(
-            self,
-            *,
-            exp_format: str = 'json',
-            record_ids: Optional[list[str]] = None,
-            fields: Optional[list[str]] = None,
-            forms: Optional[list[str]] = None,
-            events: Optional[list[str]] = None,
-            filters: Optional[str] = None) -> List[Dict[str, str]] | str:
-        """Export records from the REDCap project.
-
-        Args:
-            exp_format: Export format, defaults to 'json'
-            record_ids (Optional): List of record IDs to be exported
-            fields (Optional): List of fields to be included
-            forms (Optional): List of forms to be included
-            events (Optional): List of events to be included
-            filters (Optional) : Filter logic as a string (e.g. [age]>30)
-
-        Returns:
-            The list of records (JSON objects) or
-            a CSV text string depending on exp_format.
-
-        Raises:
-          REDCapConnectionError if the response has an error.
-        """
-
-        data = {
-            'content': 'record',
-            'action': 'export',
-            'returnFormat': 'json'
-        }
-
-        # If set of record ids specified, export only those records.
-        if record_ids:
-            data['records'] = ','.join(record_ids)
-
-        # If set of fields specified, export records only those fields.
-        if fields:
-            data['fields'] = ','.join(fields)
-
-        # If set of forms specified, export records only from those forms.
-        if forms:
-            data['forms'] = ','.join(forms)
-
-        # If set of events specified, export records only for those events.
-        if events:
-            data['events'] = ','.join(events)
-
-        # If any filters specified, export only matching records.
-        if filters:
-            data['filterLogic'] = ','.join(filters)
-
-        message = 'failed to export records'
-        if exp_format == 'json':
-            return self.request_json_value(data=data, message=message)
-
-        return self.request_text_value(data=data,
-                                       result_format=exp_format,
-                                       message=message)
-
-    def export_field_names(self) -> List[Dict[str, str]]:
+    def __export_field_names(self) -> List[Dict[str, str]]:
         """Get the export field names for the project variables.
 
         Returns:
@@ -331,7 +219,7 @@ class REDCapConnection:
 
         return self.request_json_value(data=data, message=message)
 
-    def export_project_info(self) -> Dict[str, Any]:
+    def __export_project_info(self) -> Dict[str, Any]:
         """Export the basic attributes of the project.
 
         Returns:
@@ -343,79 +231,6 @@ class REDCapConnection:
 
         message = "exporting project info"
         data = {'content': 'project'}
-
-        return self.request_json_value(data=data, message=message)
-
-    def export_instruments(self) -> List[Dict[str, str]]:
-        """Export the list of instruments in the project.
-
-        Returns:
-            List containing the name and label for each instrument
-
-        Raises:
-          REDCapConnectionError if the response has an error
-        """
-
-        message = "exporting list of forms"
-        data = {'content': 'instrument'}
-
-        return self.request_json_value(data=data, message=message)
-
-    def export_user_roles(self) -> List[Dict[str, Any]]:
-        """Export user roles defined in the project.
-
-        Returns:
-            List of user role dicts specifying permissions for each role
-
-        Raises:
-          REDCapConnectionError if the response has an error
-        """
-        message = "exporting user roles"
-        data = {'content': 'userRole'}
-
-        return self.request_json_value(data=data, message=message)
-
-    def assign_user_role(self, username: str, role: str) -> int:
-        """Assign given user to a user role in REDCap project.
-
-        Args:
-            username: REDCap username
-            role: Unique REDCap generated role name (not role label)
-
-        Returns:
-            Number of User-Role assignments added or updated
-
-        Raises:
-          REDCapConnectionError if the response has an error
-        """
-
-        message = f"assigning user {username} to role {role}"
-        info = {"username": username, "unique_role_name": role}
-        data = json.dumps([info])
-        data = {'content': 'userRoleMapping', 'action': 'import', 'data': data}
-
-        return self.request_json_value(data=data, message=message)
-
-    def add_user(self, user_info: Dict[str, Any]) -> int:
-        """Import a new user into a project and set user privileges, or update
-        the privileges of an existing user in the project.
-
-        Args:
-            user_info: User permissions for the project
-
-        Returns:
-            int: Number of users added or updated
-
-        Raises:
-          REDCapConnectionError if the response has an error
-        """
-
-        message = f"adding user {user_info['username']}"
-        info = json.dumps([user_info])
-        data = {
-            'content': 'user',
-            'data': info,
-        }
 
         return self.request_json_value(data=data, message=message)
