--- common/src/python/redcap/redcap_permissions.py
@@ -1,121 +1,0 @@
-"""Methods to define/set user permissions in REDCap projects."""
-
-import logging
-from typing import Any, Dict, List, Optional
-
-from redcap.redcap_connection import REDCapConnection, REDCapConnectionError
-
-ENROLLMENT_MODULE = 'ENROLLV1'
-NACC_TECH_ROLE = 'NACC-TECH-ROLE'
-NACC_STAFF_ROLE = 'NACC-STAFF-ROLE'
-CENTER_USER_ROLE = 'CENTER-USER-ROLE'
-GEARBOT_USER_ID = 'nacc-flywheel-gear@uw.edu'
-
-log = logging.getLogger()
-
-
-def get_nacc_developer_permissions(
-        *,
-        username: str,
-        expiration: Optional[str] = None,
-        forms_list: Optional[List[Dict[str, str]]] = None) -> Dict[str, Any]:
-    """Permissions for a NACC user who has developer privilleges for a project.
-
-    Args:
-        username: REDCap username
-        expiration (optional): permission expiration date
-        forms_list (optional): list of forms in the project
-
-    Returns:
-        Dict[str, Any]: user permissions directory
-    """
-
-    forms = {}  # Form rights
-    forms_export = {}  # Data export rights
-
-    # Need to set permissions for each form in the project
-    if forms_list:
-        for form in forms_list:
-            form_name = form['instrument_name']
-            forms[form_name] = 1  # View and Edit
-            forms_export[form_name] = 1  # Full Data Set
-
-    permissions = {
-        "username": username,
-        "expiration": expiration,
-        "design": 1,
-        "alerts": 1,
-        "user_rights": 1,
-        "data_access_groups": 1,
-        "data_export": 1,
-        "reports": 1,
-        "stats_and_charts": 1,
-        "manage_survey_participants": 1,
-        "calendar": 1,
-        "data_import_tool": 1,
-        "data_comparison_tool": 1,
-        "logging": 1,
-        "file_repository": 1,
-        "data_quality_create": 1,
-        "data_quality_execute": 1,
-        "api_export": 1,
-        "api_import": 1,
-        "record_create": 1,
-        "record_rename": 1,
-        "record_delete": 1,
-        "forms": forms,
-        "forms_export": forms_export
-    }
-
-    return permissions
-
-
-def assign_update_user_role_in_redcap(username: str, role_label: str,
-                                      redcap_con: REDCapConnection) -> bool:
-    """Assign or update user permissions in the REDCap project.
-
-    Args:
-        username: REDCap user name
-        role_label: REDCap user role to be assigned to the user
-        redcap_con: API connection for the REDCap project
-    """
-
-    try:
-        roles = redcap_con.export_user_roles()
-        role_name = None
-        for role in roles:
-            if role['role_label'] == role_label:
-                role_name = role['unique_role_name']
-                break
-
-        if not role_name:
-            log.error('User role %s does not exist in REDCap project %s',
-                      role_label, redcap_con.title)
-            return False
-
-        redcap_con.assign_user_role(username, role_name)
-    except REDCapConnectionError as error:
-        log.error(
-            'Failed to assign/update permissions for user %s in REDCap project %s - %s',
-            username, redcap_con.title, error)
-        return False
-
-    return True
-
-
-def add_gearbot_user_to_project(redcap_con: REDCapConnection):
-    """Add nacc gearbot user to the specified project.
-
-    Args:
-        redcap_con: API connection for the REDCap project
-
-    Raises:
-        REDCapConnectionError if the response has an error.
-    """
-
-    if not assign_update_user_role_in_redcap(GEARBOT_USER_ID, NACC_TECH_ROLE,
-                                             redcap_con):
-        forms = redcap_con.export_instruments()
-        gearbot_user = get_nacc_developer_permissions(username=GEARBOT_USER_ID,
-                                                      forms_list=forms)
-        redcap_con.add_user(gearbot_user)
